+++ test/uk/gov/hmrc/decisionservice/ruleengine/SectionEmptyValuesMatchingSpec.scala
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.decisionservice
+package uk.gov.hmrc.decisionservice.ruleengine
 
 import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.{BeforeAndAfterEach, Inspectors, LoneElement}
 import uk.gov.hmrc.decisionservice.model._
 import uk.gov.hmrc.decisionservice.model.rules.{SectionRuleSet, _}
-import uk.gov.hmrc.decisionservice.ruleengine.FactMatcherInstance
 import uk.gov.hmrc.play.test.UnitSpec
 
 class SectionEmptyValuesMatchingSpec extends UnitSpec with BeforeAndAfterEach with ScalaFutures with LoneElement with Inspectors with IntegrationPatience {
