+++ test/uk/gov/hmrc/decisionservice/services/TablesNonFinalDecisionSpec.scala
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.decisionservice
+package uk.gov.hmrc.decisionservice.services
 
 import uk.gov.hmrc.decisionservice.model.rules.{>>>, Facts}
 import uk.gov.hmrc.decisionservice.ruleengine.RulesFileMetaData
-import uk.gov.hmrc.decisionservice.service.DecisionService
 import uk.gov.hmrc.play.test.UnitSpec
 
 class TablesNonFinalDecisionSpec extends UnitSpec {
